op_o2 -> op_o1 (remove old op_o1)
