+++ my_pyscf/lassi/op_o1/rdm.py
@@ -3,9 +3,9 @@ import numpy as np
 from scipy import linalg
 from pyscf import lib
 from pyscf.lib import logger
-from mrh.my_pyscf.lassi.op_o2 import frag
-from mrh.my_pyscf.lassi.op_o2 import stdm
-from mrh.my_pyscf.lassi.op_o2.utilities import *
+from mrh.my_pyscf.lassi.op_o1 import frag
+from mrh.my_pyscf.lassi.op_o1 import stdm
+from mrh.my_pyscf.lassi.op_o1.utilities import *
 
 # C interface
 import ctypes
