+++ my_pyscf/lassi/sitools.py
@@ -4,7 +4,7 @@ from scipy import linalg
 from mrh.my_pyscf.mcscf.lasci import get_space_info
 from mrh.my_pyscf.lassi.citools import get_lroots, get_rootaddr_fragaddr
 from mrh.my_pyscf.lassi.lassi import root_make_rdm12s, LASSI, ham_2q
-from mrh.my_pyscf.lassi.op_o1 import fermion_spin_shuffle
+from mrh.my_pyscf.lassi.op_o1.utilities import fermion_spin_shuffle
 
 def decompose_sivec_by_rootspace (las, si, ci=None):
     '''Decompose a set of LASSI vectors as
