+++ tests/lassi/test_c2h4n4.py
@@ -158,7 +158,7 @@ class KnownValues(unittest.TestCase):
             self.assertLess (np.amax (np.abs (errvec)), 1e-8)
 
     def test_lassis (self):
-        for opt in (0,1,2):
+        for opt in (0,1):
             with self.subTest (opt=opt):
                 from mrh.my_pyscf.lassi.lassis import LASSIS
                 las1 = LASSCF (las._scf, (4,4), (4,4), spin_sub=(1,1))
@@ -178,7 +178,7 @@ class KnownValues(unittest.TestCase):
         las1 = LASSCF (mf, (5,5), ((3,2),(2,3)), spin_sub=(2,2))
         mo_coeff = las1.localize_init_guess ((list (range (5)), list (range (5,10))))
         las1.kernel (mo_coeff)
-        for opt in (0,1,2):
+        for opt in (0,1):
             with self.subTest (opt=opt):
                 lsis = LASSIS (las1).run (opt=opt)
                 self.assertAlmostEqual (lsis.e_roots[0], -295.5210783894406, 7)
